# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: National LMIS Ledger System
  description: Decoupled Modular Monolith for Supply Chain Accounting
  version: 1.0.0
paths:
...
  /api/adapter/admin/contracts:
    get:
      tags:
      - Adapter Admin
      summary: List Contracts
      operationId: list_contracts_api_adapter_admin_contracts_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  additionalProperties: true
                  type: object
                type: array
                title: Response List Contracts Api Adapter Admin Contracts Get
      security:
      - HTTPBearer: []
    post:
      tags:
      - Adapter Admin
      summary: Create Contract
      operationId: create_contract_api_adapter_admin_contracts_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContractCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/adapter/admin/contracts/{contract_id}/versions/{version}/activate:
    post:
      tags:
      - Adapter Admin
      summary: Activate Contract
      description: Implements the 'Atomic Flip' rule.
      operationId: activate_contract_api_adapter_admin_contracts__contract_id__versions__version__activate_post
      security:
      - HTTPBearer: []
      parameters:
      - name: contract_id
        in: path
        required: true
        schema:
          type: string
          title: Contract Id
      - name: version
        in: path
        required: true
        schema:
          type: string
          title: Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/adapter/admin/crosswalks:
    get:
      tags:
      - Adapter Admin
      summary: List Crosswalks
      operationId: list_crosswalks_api_adapter_admin_crosswalks_get
      security:
      - HTTPBearer: []
      parameters:
      - name: namespace
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Namespace
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response List Crosswalks Api Adapter Admin Crosswalks Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Adapter Admin
      summary: Create Crosswalk
      operationId: create_crosswalk_api_adapter_admin_crosswalks_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrosswalkCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/adapter/admin/dlq:
    get:
      tags:
      - Adapter Admin
      summary: Get Dlq
      operationId: get_dlq_api_adapter_admin_dlq_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status_filter
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          default: UNRESOLVED
          title: Status Filter
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  additionalProperties: true
                title: Response Get Dlq Api Adapter Admin Dlq Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/adapter/admin/dlq/{id}/retry:
    post:
      tags:
      - Adapter Admin
      summary: Retry Dlq
      operationId: retry_dlq_api_adapter_admin_dlq__id__retry_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ContractCreate:
      properties:
        id:
          type: string
          title: Id
        version:
          type: string
          title: Version
        dsl_config:
          additionalProperties: true
          type: object
          title: Dsl Config
      type: object
      required:
      - id
      - version
      - dsl_config
      title: ContractCreate
    CrosswalkCreate:
      properties:
        namespace:
          type: string
          title: Namespace
        source_value:
          type: string
          title: Source Value
        internal_id:
          type: string
          title: Internal Id
        metadata_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata Json
      type: object
      required:
      - namespace
      - source_value
      - internal_id
      title: CrosswalkCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer